+++ integration/apidefinition_test.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/TykTechnologies/tyk-operator/api/model"
 	"github.com/TykTechnologies/tyk-operator/api/v1alpha1"
+	"github.com/google/uuid"
 	"github.com/matryer/is"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -695,4 +696,54 @@ Q1+khpfxP9x1H+mMlUWBgYPq7jG5ceTbltIoF/sUQPNR+yKIBSnuiISXFHO9HEnk
 		Feature()
 
 	testenv.Test(t, adCreate)
+}
+
+func TestAPIDefinition_GraphQL_ExecutionMode(t *testing.T) {
+	createAPI := features.New("Create GraphQL API").
+		Assess("validate_executionMode", func(ctx context.Context, t *testing.T, c *envconf.Config) context.Context {
+			testNS := ctx.Value(ctxNSKey).(string) //nolint:errcheck
+			is := is.New(t)
+
+			tests := map[string]struct {
+				ExecutionMode string
+				ReturnErr     bool
+			}{
+				"invalid execution mode": {
+					ExecutionMode: "invalid",
+					ReturnErr:     true,
+				},
+				"empty execution mode": {
+					ExecutionMode: "",
+					ReturnErr:     true,
+				},
+				"valid execution engine mode": {
+					ExecutionMode: "executionEngine",
+					ReturnErr:     false,
+				},
+				"valid proxy only mode": {
+					ExecutionMode: "proxyOnly",
+					ReturnErr:     false,
+				},
+			}
+
+			for n, tc := range tests {
+				t.Run(n, func(t *testing.T) {
+					_, err := createTestAPIDef(ctx, testNS, func(ad *v1alpha1.ApiDefinition) {
+						ad.Name = fmt.Sprintf("%s-%s", ad.Name, uuid.New().String())
+						ad.Spec.Name = ad.Name
+						ad.Spec.GraphQL = &model.GraphQLConfig{
+							Enabled:       true,
+							ExecutionMode: model.GraphQLExecutionMode(tc.ExecutionMode),
+						}
+					}, c)
+
+					t.Log("Error=", err)
+					is.Equal(tc.ReturnErr, err != nil)
+				})
+			}
+
+			return ctx
+		}).Feature()
+
+	testenv.Test(t, createAPI)
 }